Fix ownership problem in TMVAEvaluator
The TMVAEvaluator was meant to own the GBRForest if the TMVAEvaluator
made the instance itself else it was not supposed to own it if it
was given a GBRForest externally. However, for both cases the code
used an std::unique_ptr and attempted to call 'release' in the case
of non-ownership. The problem was multiple calls to initializeGBRForest
still caused the code to delete a GBRForest which was not supposed to be
owned. The code was changed to use an std::shared_ptr where we use a custom
deallocator in the case of non-ownership (the deallocator does nothing).
This problem was found in the IB RelVals.
(cherry picked from commit 3600253af5ada0564d2f9af11ee217d7429ceb8b)